Migrations updated and service added

+++ Hairdresser/Controllers/AdminController.cs
@@ -10,11 +10,12 @@ namespace Hairdresser.Controllers
     public class AdminController : Controller
     {
         private readonly DbContext1 dbContext;
-
+  
         public AdminController(DbContext1 databaseContext)
         {
             dbContext = databaseContext;
         }
+        
         public IActionResult Index()
         {
             var employees = dbContext.Calisan.ToList();
@@ -32,6 +33,7 @@ namespace Hairdresser.Controllers
         }
         public IActionResult AddEmployee(Calisan e)
         {
+            
             if (ModelState.IsValid)
             {
                 if (dbContext.Users.Any(x => x.Username.ToLower() == e.Name.ToLower()))
@@ -46,6 +48,10 @@ namespace Hairdresser.Controllers
                     Role = "employee"
                 };
                 dbContext.Users.Add(user);
+                dbContext.SaveChanges();
+                e.User = user;
+                e.UserId = user.Id;
+                user.Calisan = e;
                 dbContext.Calisan.Add(e);
                 int affectedRowCount = dbContext.SaveChanges();
 
@@ -70,21 +76,60 @@ namespace Hairdresser.Controllers
                 return RedirectToAction("Team");
             }
             var employees = dbContext.Calisan.Find(id);
+            var user = dbContext.Users.Find(employees.UserId);
             if (employees is null)
             {
                 TempData["msj"] = "Employee could not find";
                 return RedirectToAction("Team");
             }
-            User user = dbContext.Users.SingleOrDefault(x => x.Username.ToLower() == employees.Name.ToLower() && x.Password == employees.Password);
+            if(user is null)
+            {
+                TempData["msj"] = "User could not find";
+                return RedirectToAction("Team");
+            }
+            //User user = dbContext.Users.SingleOrDefault(x => x.Username.ToLower() == employees.Name.ToLower() && x.Password == employees.Password);
             dbContext.Users.Remove(user);
             dbContext.Calisan.Remove(employees);
             dbContext.SaveChanges();
             TempData["msj"] = employees.Name + " named employee deleted.";
             return RedirectToAction("Team");
         }
+
+        public void HepsiniSil()
+        {
+            var employee = dbContext.Calisan.ToList();
+            dbContext.Calisan.RemoveRange(employee);
+            dbContext.SaveChanges();
+            var users = dbContext.Users.ToList();
+            dbContext.Users.RemoveRange(users);
+            dbContext.SaveChanges();
+        }
+        public IActionResult GetMembers()
+        {
+            List<Calisan> memberListDB = dbContext.Calisan.ToList();
+            List<Calisan> memberList = new List<Calisan>();
+
+            foreach (Calisan Member in memberListDB)
+            {
+                memberList.Add(new Calisan
+                {
+                    Id = Member.Id,
+                    Name = Member.Name,
+                    Surname = Member.Surname,
+                    Salary = Member.Salary,
+                    Position = Member.Position,
+                    Email = Member.Email,
+                    Password = Member.Password,
+                    User = Member.User,
+                    UserId = Member.UserId,
+                    PhoneNumber = Member.PhoneNumber
+                });
+            }
+            return View(memberList);
+        }
         public IActionResult EmployeeEdit(int? id)
         {
-            if (id is null)
+            if(id is null)
             {
                 TempData["msj"] = "Please enter valid data";
                 return RedirectToAction("EmployeeEdit");
@@ -95,6 +140,14 @@ namespace Hairdresser.Controllers
                 TempData["msj"] = "Employee could not find";
                 return RedirectToAction("EmployeeEdit");
             }
+            if (employee.UserId is null)
+            {
+                TempData["msj"] = "Please enter valid data";
+                return RedirectToAction("EmployeeEdit");
+            }
+            var user = dbContext.Users.Find(employee.UserId);
+            dbContext.Users.Remove(user);
+            dbContext.SaveChanges();
             return View(employee);
         }
         [HttpPost]
@@ -113,14 +166,22 @@ namespace Hairdresser.Controllers
 
             if (ModelState.IsValid)
             {
-                User user = dbContext.Users.SingleOrDefault(x => x.Username.ToLower() == e.Name.ToLower() && x.Password == e.Password);
-
-                dbContext.Users.Update(user);
+                User user = new()
+                {
+                    Username = e.Name,
+                    Password = e.Password,
+                    Role = e.Position
+                };
+                dbContext.Users.Add(user);
+                dbContext.SaveChanges();
+                e.User = user;
+                e.UserId = user.Id;
+                user.Calisan = e;
                 dbContext.Calisan.Update(e);
                 
                 dbContext.SaveChanges();
                 TempData["msj"] = e.Name + " named employee updated.";
-                return RedirectToAction("EmployeeEdit");
+                return RedirectToAction("Team");
             }
             TempData["msj"] = "Update failed!";
             return RedirectToAction("EmployeeEdit");
@@ -141,5 +202,97 @@ namespace Hairdresser.Controllers
             }
             return View(employees);
         }
+
+        public IActionResult Service()
+        {
+            var services = dbContext.Services.ToList();
+            return View(services);
+        }
+        public IActionResult AdminServiceAdd()
+        {
+            return View();
+        }
+        public IActionResult AddService(Servis s)
+        {
+            if (ModelState.IsValid)
+            {
+                if (dbContext.Services.Any(x => x.Name.ToLower() == s.Name.ToLower()))
+                {
+                    ModelState.AddModelError(nameof(s.Name), "Service already exists");
+                    View(s);
+                }
+                dbContext.Services.Add(s);
+                int affectedRowCount = dbContext.SaveChanges();
+
+                if (affectedRowCount == 0)
+                {
+                    ModelState.AddModelError("", "User can not be added");
+                }
+                else
+                {
+                    return RedirectToAction(nameof(Service));
+                }
+            }
+            TempData["msj"] = "Please enter a valid data";
+            return RedirectToAction("AdminServiceAdd");
+        }
+        public IActionResult DeleteService(int? id)
+        {
+            if (id is null)
+            {
+                TempData["msj"] = "Please enter valid data";
+                return RedirectToAction("Service");
+            }
+            var service = dbContext.Services.Find(id);
+            if (service is null)
+            {
+                TempData["msj"] = "Employee could not find";
+                return RedirectToAction("Team");
+            }
+            dbContext.Services.Remove(service);
+            dbContext.SaveChanges();
+            TempData["msj"] = service.Name + " named service deleted.";
+            return RedirectToAction("Service");
+        }
+        public IActionResult EditService(int? id)
+        {
+            if (id is null)
+            {
+                TempData["msj"] = "Please enter valid data";
+                return RedirectToAction("EditService");
+            }
+            var service = dbContext.Services.Find(id);
+            if (service is null)
+            {
+                TempData["msj"] = "Employee could not find";
+                return RedirectToAction("EditService");
+            }
+            return View(service);
+        }
+        [HttpPost]
+        public IActionResult EditService(int? id,Servis s)
+        {
+            if (id is null)
+            {
+                TempData["msj"] = "Please enter valid data";
+                return RedirectToAction("EditService");
+            }
+            if (id != s.Id)
+            {
+                TempData["msj"] = "Id does not exist.";
+                return RedirectToAction("EditService");
+            }
+
+            if (ModelState.IsValid)
+            {
+                dbContext.Services.Update(s);
+                dbContext.SaveChanges();
+                TempData["msj"] = s.Name + " named employee updated.";
+                return RedirectToAction("Service");
+            }
+            TempData["msj"] = "Update failed!";
+            return RedirectToAction("EditService");
+        }
+
     }
 }
